fix: python scripts (rename scripts -> kakarot_scripts)
