+++ .env.example
@@ -1,5 +1,5 @@
 # A Github token to download the CI artifacts
-# Used for fetching resources in the scripts
+# Used for fetching resources in the kakarot_scripts
 GITHUB_TOKEN=
 
 # Starkent Network
@@ -8,7 +8,7 @@ STARKNET_NETWORK=katana
 # An infura RPC key for mainnet/testnet
 INFURA_KEY=
 
-# Account used by default in all scripts
+# Account used by default in all kakarot_scripts
 # Prefixed accounts {NETWORK}_ will take precedeance over the default one
 MADARA_ACCOUNT_ADDRESS=0x3
 MADARA_PRIVATE_KEY=0x00c1cf1490de1352865301bb8705143f3ef938f97fdf892f1090dcb5ac7bcd1d
@@ -28,7 +28,7 @@ KATANA_PRIVATE_KEY=0x2bbf4f9fd0bbb2e60b0316c1fe0b76cf7a4d0198bd493ced9b8df2a3a24
 ACCOUNT_ADDRESS=
 PRIVATE_KEY=
 
-# An EVM private to define a default EOA for EVM related scripts
+# An EVM private to define a default EOA for EVM related kakarot_scripts
 # This default value is Anvil first account private key
 EVM_PRIVATE_KEY=0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
