+++ .github/workflows/ci.yml
@@ -204,7 +204,7 @@ jobs:
           make ef-test-v0 | tee test_v0.out
           set +o pipefail
       - name: Retrieve ef-tests execution resources
-        run: python scripts/compute_resources.py
+        run: python kakarot_scripts/compute_resources.py
         env:
           KAKAROT_VERSION: v0
       - name: upload resources
