+++ docs/general/decode_a_cairo_trace.md
@@ -39,8 +39,8 @@ this cryptic message to map location in the original cairo program.
   - by default in the `compile_kakarot.py` script, the `--no_debug_info` is
     applied when the target network is not a devnet (e.g katana)
   - to use the tooling as is, just run
-    `STARKNET_NETWORK=starknet-devnet python scripts/compile_kakarot.py` or
-    `make build` if you set katana as your `STARKNET_NETWORK` in `.env`
+    `STARKNET_NETWORK=starknet-devnet python kakarot_scripts/compile_kakarot.py`
+    or `make build` if you set katana as your `STARKNET_NETWORK` in `.env`
 
 - Then, use this simple snippet as a standalone Python file
 
@@ -48,7 +48,7 @@ this cryptic message to map location in the original cairo program.
   # %% Imports
   import json
 
-  from scripts.utils.starknet import get_artifact
+  from kakarot_scripts.utils.starknet import get_artifact
   from starkware.cairo.lang.compiler.program import Program
   from starkware.cairo.lang.vm.reconstruct_traceback import reconstruct_traceback
 
@@ -68,12 +68,12 @@ this cryptic message to map location in the original cairo program.
 And tada! You should now be able to understand the reason your cairo program
 failed with more clarity.
 
-## Extended scripts
+## Extended kakarot_scripts
 
 As it's not always obvious which contract triggered the error (Kakarot, account,
 EVM for the bytecode tests), this extended snippets proved to be more robust as
-a general purpose debug scripts, to be run both from the ef-test and the kakarot
-repo.
+a general purpose debug kakarot_scripts, to be run both from the ef-test and the
+kakarot repo.
 
 It basically loads by default all the possible contract sources, looking for
 them in appropriate directories, and then run the debug tool for each of them.
