+++ kakarot_scripts/constants.py
