+++ kakarot_scripts/ef_tests/fetch.py
