+++ kakarot_scripts/ef_tests/resources.py
