+++ kakarot_scripts/utils/erc20.json
