+++ tests/end_to_end/test_kakarot.py
@@ -6,7 +6,7 @@ import pytest_asyncio
 from starknet_py.contract import Contract
 from starknet_py.net.full_node_client import FullNodeClient
 
-from scripts.utils.starknet import wait_for_transaction
+from kakarot_scripts.utils.starknet import wait_for_transaction
 from tests.end_to_end.bytecodes import test_cases
 from tests.utils.constants import PRE_FUND_AMOUNT, TRANSACTION_GAS_LIMIT
 from tests.utils.helpers import (
@@ -38,7 +38,10 @@ async def other():
     """
     Just another Starknet contract.
     """
-    from scripts.utils.starknet import deploy_starknet_account, get_starknet_account
+    from kakarot_scripts.utils.starknet import (
+        deploy_starknet_account,
+        get_starknet_account,
+    )
 
     account_info = await deploy_starknet_account()
     return await get_starknet_account(account_info["address"])
@@ -49,7 +52,7 @@ async def class_hashes():
     """
     All declared class hashes.
     """
-    from scripts.utils.starknet import get_declarations
+    from kakarot_scripts.utils.starknet import get_declarations
 
     return get_declarations()
 
@@ -59,7 +62,7 @@ async def origin(evm: Contract, addresses):
     """
     Deploys the origin's Starknet contract to the correct address and funds it.
     """
-    from scripts.utils.starknet import fund_address
+    from kakarot_scripts.utils.starknet import fund_address
 
     evm_address = int(addresses[0].address, 16)
     sn_address = (
