+++ tests/utils/helpers.py
@@ -11,7 +11,7 @@ from eth_keys import keys
 from eth_utils import decode_hex, keccak, to_checksum_address
 from starkware.starknet.public.abi import get_storage_var_address
 
-from scripts.constants import NETWORK
+from kakarot_scripts.constants import NETWORK
 from tests.utils.uint256 import int_to_uint256
 
 PERMIT_TYPEHASH = keccak(
